Connection to AAVE on test fork is functional, and ready to take on new eToken algo writes/tests.  Pushing to move codebase to new computer.

+++ contracts/peripheries/interfaces/IeaEons.sol
@@ -5,7 +5,7 @@ import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 import '../utilities/MinterRole.sol';
 
 interface IeaEons is IERC20 {
-  function mint(address recepient, uint amount) external;
+  function mint(address recepient, uint256 amount) external;
   function addMinter(address minter) external;
   function burn(address from, uint256 amount) external;
   function eBalanceOf(address user) external returns(uint256);
